debian: fix dh-3pc usage
There was a trivial bug in dh-3pc, which most of its users relied on,
because there also was a conceptual bug in dh-3pc, which otherwise
would have made it unusable for common multi-binary packages.  Both
were fixed in dh-3pc 0.58, so we have to adapt our old broken usage.